pyproject.toml
[project]
name = "resumeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.0",
    "google-genai>=1.12.1",
    "ipykernel>=6.29.5",
    "latex>=0.7.0",
    "markdown>=3.8",
    "markdown-pdf>=1.7",
    "pdflatex>=0.1.3",
    "pylatex>=1.4.2",
    "pypdf>=5.4.0",
    "requests>=2.32.3",
]
